ai-service/requirements.txt
flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.4
pymongo==4.10.1
python-dotenv==1.0.1
